Support Zeitwerk and Redmine5/Rails6
